+++ backend/budget/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from budget.models import Budget
+
+admin.site.register(Budget)