+++ backend/familybudget/settings/base.py
@@ -39,6 +39,8 @@ INSTALLED_APPS = [
     "rest_framework",
     "common",
     "users",
+    "transaction",
+    "budget",
 ]
 
 MIDDLEWARE = [