+++ docs/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.3.5]] - 2022-03-24
+
+###### _([diff: 0.3.4-0.3.5])_
+
+### Added
+
+- Separate Apps for [`budget`](../backend/budget/) and [`transaction`](../backend/transaction/)
+- Serializers for [`Income`](../backend/transaction/serializer.py), [`Expense`](../backend/transaction/serializer.py), [`User`](../backend/users/serializer.py) and [`Budget`](../backend/budget/serializer.py)
+
 ## [[0.3.4]] - 2022-03-24
 
 ###### _([diff: 0.3.3-0.3.4])_
@@ -170,6 +179,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.3.4-0.3.5]: https://github.com/Luzkan/FamilyBudget/compare/0.3.4...0.3.5
 [diff: 0.3.3-0.3.4]: https://github.com/Luzkan/FamilyBudget/compare/0.3.3...0.3.4
 [diff: 0.3.2-0.3.3]: https://github.com/Luzkan/FamilyBudget/compare/0.3.2...0.3.3
 [diff: 0.3.1-0.3.2]: https://github.com/Luzkan/FamilyBudget/compare/0.3.1...0.3.2
@@ -180,6 +190,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.3.5]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.5
 [0.3.4]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.4
 [0.3.3]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.3
 [0.3.2]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.2